Add single test for Base#mail method
This commit adds a test to ensure that a new Message instance is created
when the `Base#mail` method is called. In order to do this, it was
necessary to add a new private method to the Base class as the
`instance_variable_names` method called is a Rails-only method [1] and
so would have raised an exception in a non-Rails application.

[1]: https://apidock.com/rails/v5.2.3/Object/instance_variable_names